Implemented the local and remote party to make it easier to keep track of who's information we are dealing.
The Caller-ID and connected-line ID are just roles the local and remote party
play while a call is being established.  Which roll depends upon the direction
of the call.
Outgoing party info is to identify the local party to the other end.
   (Caller-ID for originated or connected-line for answered calls.)
Incoming party info is to identify the remote party to us.
   (Caller-ID for answered or connected-line for originated calls.)