Commit checkpoint for work in progress.
*  Rearange code to make some pri_facility.c functions available to
pri_cc.c.
*  Some progress on PTMP call completion.
*  Make q931_encode_channel() handle the cis_call flag like the held_call
flag.  Asterisk will need to deal with PRI events coming in with the
cis_call flag set in the channel parameter.  Asterisk should generally
treat subcmds from these events as coming in without a call.
*  q931_fill_facility_event() now sets the PRI_EVENT_FACILITY call
parameter to NULL if the call reference is the dummy call reference.