Need to block the stop alerting event when we are the one doing the CC recall.
Added CC_STATE_WAIT_CALLBACK and CC_STATE_CALLBACK to the PTMP monitor FSM
to be more precise in handling the CC events.  Could not get away with the
simplified FSM.  Bad things happened in the upper layer when the
PRI_SUBCMD_CC_STOP_ALERTING event was not blocked when we are the one
doing the CC recall call.