Branch for issue 17865
BRI NT ptmp: Active channels not cleard when i/f goes down

If the connection to the terminal is lost, while there are open channels
on the interface, red alarm is reported, but the open channels are never
cleared.

Additionally, if you manually try to channel request hangup, Asterisk
crashes.