Fix T309 behaviour.
* Made T309 processing use the correct call pool.  We were looking in the
wrong place for the call pool.

* Made always start T309 when a call is in Q931_CALL_STATE_ACTIVE since
the UP event assumes that the active timer is T309.

* Removed code that checked for the disabled T309 exception.  The user
could never disable the T309 timer once it is enabled and it is now
enabled by default.  The timer value can never be negative anymore since
it is default enabled and the set routine rejects setting any negative
value.

* Made timeout routines clear the retranstimer id value if the timout
routine was using that member to hold the id.